--- srcs/clients.cpp
@@ -1,21 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   clients.cpp                                        :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: abberkac <abberkac@student.42.fr>          +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/03/21 01:39:00 by abberkac          #+#    #+#             */
-/*   Updated: 2024/03/30 04:43:17 by abberkac         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "../includes/clients.hpp"
-
-Clients::Clients() {
-    return;
-}
-
-Clients::~Clients() {
-    return;
-}
