--- srcs/clients.hpp
@@ -1,25 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   clients.hpp                                        :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: abberkac <abberkac@student.42.fr>          +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/03/21 01:37:16 by abberkac          #+#    #+#             */
-/*   Updated: 2024/03/31 01:53:18 by abberkac         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#ifndef CLIENTS_HPP
-#define CLIENTS_HPP
-
-class Clients {
-  private:
-    // int _sd;
-    
-    public:
-        Clients();
-        ~Clients();
-};
-
-#endif
